rename DescribeCommand to GenerateRectorOverviewCommand
