+++ src/Rector/Dynamic/ClassReplacerRector.php
@@ -31,7 +31,7 @@ final class ClassReplacerRector extends AbstractRector
      */
     public function getDefinition(): RectorDefinition
     {
-        return new RectorDefinition('[Dynamic] Replaces defined classes by new ones.', [
+        return new RectorDefinition('Replaces defined classes by new ones.', [
             new CodeSample('$value = new SomeOldClass;', '$value = new SomeNewClass;'),
         ]);
     }
