+++ src/Rector/MagicDisclosure/GetAndSetToMethodCallRector.php
@@ -62,7 +62,7 @@ final class GetAndSetToMethodCallRector extends AbstractRector
 
     public function getDefinition(): RectorDefinition
     {
-        return new RectorDefinition('[Dynamic] Turns defined `__get`/`__set` to specific method calls.', [
+        return new RectorDefinition('Turns defined `__get`/`__set` to specific method calls.', [
             new CodeSample(
                 '$someService = $container->someService;',
                 '$someService = $container->getService("someService");'
